update queue styles
