+++ app/components/Homepage/Queue/Queue.module.scss
@@ -1,15 +1,16 @@
 @use 'styles' as *;
 
-.layout {
-  display: grid;
-  grid-template-rows: 5rem auto;
-  height: 100%;
-}
-
 .title {
   @extend %text-style-display-md-semibold;
 
   color: $color-gray-50;
+}
+
+.titlerow {
+  align-items: center;
+  display: flex;
+  justify-content: space-between;
+  width: 100%;
 }
 
 .playerSelect {
@@ -24,27 +25,8 @@
 
 .list {
   color: $color-gray-200;
-  grid-column: 1 / span 2;
+  display: flex;
+  flex-direction: column;
+  height: 100%;
   overflow-y: scroll;
-}
-
-/*
- some scrollbar styles. see https://www.digitalocean.com/community/tutorials/css-scrollbars
-*/
-
-// entire scrobbar
-.list::-webkit-scrollbar {
-  width: 12px; /* width of the entire scrollbar */
-}
-
-// scrobbar background
-.list::-webkit-scrollbar-track {
-  background: $color-primary-25;
-  border-radius: $border-radius-large;
-}
-
-// scrobar "thumb"
-.list::-webkit-scrollbar-thumb {
-  background-color: $color-gray-400;
-  border-radius: $border-radius-large;
 }
