+++ app/components/Homepage/Queue/TitleRow.tsx
@@ -1,0 +1,49 @@
+import styles from './Queue.module.scss'
+import Select, { SingleValue } from 'react-select'
+import { useMemo } from 'react'
+import type { Player } from '@common/types'
+import usePlayers from 'hooks/usePlayers'
+
+interface OptionType {
+  label: string
+  value: Player
+}
+
+interface Props {
+  queue: Player[]
+  setQueue: any
+}
+
+const TitleRow = ({ queue, setQueue }: Props) => {
+  const { players } = usePlayers()
+
+  const options: OptionType[] = useMemo(
+    () =>
+      players
+        .filter(o => !queue.some(qP => qP.id === o.id))
+        .map(p => ({ label: `#${p.id} ${p.firstName} ${p.lastName}`, value: p })),
+    [players, queue]
+  )
+
+  const handleChange = (newValue: SingleValue<OptionType>) => {
+    if (newValue?.value) {
+      setQueue([...queue, newValue.value])
+    } else {
+      // eslint-disable-next-line no-console
+      console.warn('Trying to add to queue: ', newValue?.value)
+    }
+  }
+  return (
+    <div className={styles.titlerow}>
+      <h2 className={styles.title}>Queue</h2>
+      <Select
+        className={styles.playerSelect}
+        options={options}
+        onChange={handleChange}
+        placeholder="add player to queue"
+      />
+    </div>
+  )
+}
+
+export default TitleRow
