+++ app/components/Homepage/Queue/index.tsx
@@ -1,59 +1,26 @@
-import useLocalStorage from 'hooks/useLocalStorage'
-import { FunctionComponent, useMemo } from 'react'
-import { Player } from '@common/types'
+import type { Player } from '@common/types'
 import Card from '@components/utility/Card'
-import usePlayers from 'hooks/usePlayers'
-import Select, { SingleValue } from 'react-select'
 import styles from './Queue.module.scss'
 import QueueItem from './QueueItem'
+import TitleRow from './TitleRow'
+import useLocalStorage from 'hooks/useLocalStorage'
 
-interface OptionType {
-  label: string
-  value: Player
-}
-
-const Queue: FunctionComponent = () => {
-  const { players } = usePlayers()
+const Queue = () => {
   const [queue, setQueue] = useLocalStorage<Player[]>('prodeko-biliskilke-queue', [])
-
-  const options: OptionType[] = useMemo(
-    () =>
-      players
-        .filter(o => !queue.some(qP => qP.id === o.id))
-        .map(p => ({ label: `#${p.id} ${p.firstName} ${p.lastName}`, value: p })),
-    [players, queue]
-  )
-
-  const handleChange = (newValue: SingleValue<OptionType>) => {
-    if (newValue?.value) {
-      setQueue([...queue, newValue.value])
-    } else {
-      // eslint-disable-next-line no-console
-      console.warn('Trying to add to queue: ', newValue?.value)
-    }
-  }
 
   return (
     <Card rowspan="1 / 2" colspan="2 / 3">
-      <div className={styles.layout}>
-        <h2 className={styles.title}>Queue</h2>
-        <Select
-          className={styles.playerSelect}
-          options={options}
-          onChange={handleChange}
-          placeholder="add player to queue"
-        />
-        <div className={styles.list}>
-          {queue.map((player, i) => (
-            <QueueItem
-              player={player}
-              place={i + 1}
-              queue={queue}
-              setQueue={setQueue}
-              key={player.id}
-            />
-          ))}
-        </div>
+      <TitleRow queue={queue} setQueue={setQueue} />
+      <div className={styles.list}>
+        {queue.map((player, i) => (
+          <QueueItem
+            player={player}
+            place={i + 1}
+            queue={queue}
+            setQueue={setQueue}
+            key={player.id}
+          />
+        ))}
       </div>
     </Card>
   )
